tests: Uppdates tests (#98)
* tests
