+++ pipeline/Jenkinsfile
@@ -11,7 +11,7 @@ pipeline
                 choice(name: 'DRIVER', description: 'Choose browser', choices: 'chrome\nfirefox\nheadless')
 
                 // Define test path to run
-                string(name: 'TESTS_TO_RUN', defaultValue: 'all', description: 'choose test to run e.g. LT-928_edit, leave as all to run all tests')
+                string(name: 'TESTS_TO_RUN', defaultValue: '', description: 'choose test to run e.g. LT_928_edit')
 
                 // Define environment to run tests on
                 choice(name: 'ENVIRONMENT', description: 'choose what environment to run in', choices: 'staging\ndev\nuat')
