+++ ui_automation_tests/features/submit_open_application.feature
@@ -1,0 +1,105 @@
+@licence @submit @all
+Feature: I want to indicate the open licence I want
+  As a logged in exporter
+  I want to indicate the kind of licence I want
+  So that I am more likely to get the correct kind of licence or the kind of licence I would like
+
+  @LT_1091_draft
+  Scenario: Apply for a licence to draft and delete
+    Given I go to exporter homepage
+    When I click on apply for a license button
+    And I click on start button
+    And I enter in name for application and continue
+    And I select "open" application and continue
+    And I select "permanent" option and continue
+    And I select "yes" for whether I have an export licence and "123456" if I have a reference and continue
+    Then I see the application overview
+    When I delete the application
+
+  @LT_1114
+  Scenario: Submit open application
+    Given I go to exporter homepage
+    When I click on apply for a license button
+    And I click on start button
+    And I enter in name for application and continue
+    And I select "open" application and continue
+    And I select "permanent" option and continue
+    And I select "yes" for whether I have an export licence and "123456" if I have a reference and continue
+    And I click on application locations link
+    And I select "organisation" for where my goods are located
+    And I select the site at position "1"
+    And I click continue
+    And I click on the goods link from overview
+    And I click Add goods type button
+    And I add a good or good type with description "Sniper" controlled "Yes" control code "1234" incorporated "Yes" and part number "not needed"
+    Then I see my goods type added at position "1" with a description and a control code
+    When I click overview
+    Then I see my goods type added to the overview page with a description and a control code
+    When I click on the goods link from overview
+    And I click Add goods type button
+    And I click continue
+    Then I see good types error messages
+    When I add a good or good type with description "M4" controlled "Yes" control code "1234" incorporated "Yes" and part number "not needed"
+    Then I see my goods type added at position "2" with a description and a control code
+    When I click overview
+    Then I see my goods type added to the overview page with a description and a control code
+    When I click on countries
+    Then I should see a list of countries
+    When I select "Canada" from the country list
+    And I select "Poland" from the country list
+    And I select "United Kingdom" from the country list
+    And I click continue
+    Then I can see "3" countries selected on the overview page
+    When I click on number of countries on the overview page
+    Then I see "Canada" in a modal
+    And I see "Poland" in a modal
+    And I see "United Kingdom" in a modal
+    When I close the modal
+    And I submit the application
+    Then application is submitted
+    When I go to exporter homepage
+    And I click applications
+    Then I see submitted application
+
+  @LT_1092_search
+  Scenario: Search for countries
+    Given I go to exporter homepage
+    When I go to exporter homepage
+    And I click on apply for a license button
+    And I click on start button
+    And I enter in name for application and continue
+    And I select "open" application and continue
+    And I select "permanent" option and continue
+    And I select "yes" for whether I have an export licence and "123456" if I have a reference and continue
+    And I click on countries
+    And I search for country "Canada"
+    Then only "Canada" is displayed in country list
+
+  @LT_1092_error
+  Scenario: Error message when not adding countries
+    Given I go to exporter homepage
+    When I click on goods link
+    And I click add a good button
+    And I add a good or good type with description "AK47" controlled "Yes" control code "1234" incorporated "No" and part number "321"
+    And I go to exporter homepage
+    And I click on apply for a license button
+    And I click on start button
+    And I enter in name for application and continue
+    And I select "open" application and continue
+    And I select "permanent" option and continue
+    And I select "yes" for whether I have an export licence and "123456" if I have a reference and continue
+    And I click on countries
+    And I click continue
+    Then error message is "You have to pick at least one country"
+
+  @LT_1114_error
+  Scenario: Error message when not adding goods and sites information for open application
+    Given I go to exporter homepage
+    When I click on apply for a license button
+    And I click on start button
+    And I enter in name for application and continue
+    And I select "open" application and continue
+    And I select "permanent" option and continue
+    And I select "yes" for whether I have an export licence and "123456" if I have a reference and continue
+    And I click continue
+    Then I see no sites good types or countries attached error message
